requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
dateparser==1.1.8
diskcache==5.6.1
idna==3.4
llama-cpp-python==0.1.77
lxml==4.9.3
numpy==1.25.2
oauthlib==3.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
xmltodict==0.13.0
